move test connections function in its own module
