+++ psutil/tests/test_connections.py
@@ -6,6 +6,8 @@
 
 """Tests for net_connections() and Process.connections() APIs."""
 
+import contextlib
+import errno
 import os
 import socket
 import textwrap
@@ -29,7 +31,7 @@ from psutil._compat import PY3
 from psutil.tests import AF_UNIX
 from psutil.tests import bind_socket
 from psutil.tests import bind_unix_socket
-from psutil.tests import check_connection_ntuple
+from psutil.tests import check_net_address
 from psutil.tests import create_sockets
 from psutil.tests import get_free_port
 from psutil.tests import HAS_CONNECTIONS_UNIX
@@ -47,6 +49,7 @@ from psutil.tests import wait_for_file
 
 
 thisproc = psutil.Process()
+SOCK_SEQPACKET = getattr(socket, "SOCK_SEQPACKET", object())
 
 
 class Base(object):
@@ -87,7 +90,7 @@ class Base(object):
         """
         if conn is None:
             conn = self.get_conn_from_sock(sock)
-        check_connection_ntuple(conn)
+        self.check_connection_ntuple(conn)
 
         # fd, family, type
         if conn.fd != -1:
@@ -136,18 +139,108 @@ class Base(object):
         proc_cons.sort()
         self.assertEqual(proc_cons, sys_cons)
 
+    def check_connection_ntuple(self, conn):
+        """Check validity of a connection namedtuple."""
+        def check_ntuple(conn):
+            has_pid = len(conn) == 7
+            self.assertIn(len(conn), (6, 7))
+            self.assertEqual(conn[0], conn.fd)
+            self.assertEqual(conn[1], conn.family)
+            self.assertEqual(conn[2], conn.type)
+            self.assertEqual(conn[3], conn.laddr)
+            self.assertEqual(conn[4], conn.raddr)
+            self.assertEqual(conn[5], conn.status)
+            if has_pid:
+                self.assertEqual(conn[6], conn.pid)
+
+        def check_fd(conn):
+            has_fd = getattr(conn, 'fd', -1) != -1
+            if has_fd:
+                assert conn.fd >= 0, conn
+                if hasattr(socket, 'fromfd') and not WINDOWS:
+                    try:
+                        dupsock = socket.fromfd(conn.fd, conn.family,
+                                                conn.type)
+                    except (socket.error, OSError) as err:
+                        if err.args[0] != errno.EBADF:
+                            raise
+                    else:
+                        with contextlib.closing(dupsock):
+                            self.assertEqual(dupsock.family, conn.family)
+                            self.assertEqual(dupsock.type, conn.type)
+
+        def check_family(conn):
+            self.assertIn(conn.family, (AF_INET, AF_INET6, AF_UNIX))
+            # if enum is not None:
+            #     assert isinstance(conn.family, enum.IntEnum), conn
+            # else:
+            #     assert isinstance(conn.family, int), conn
+            if conn.family == AF_INET:
+                # actually try to bind the local socket; ignore IPv6
+                # sockets as their address might be represented as
+                # an IPv4-mapped-address (e.g. "::127.0.0.1")
+                # and that's rejected by bind()
+                s = socket.socket(conn.family, conn.type)
+                with contextlib.closing(s):
+                    try:
+                        s.bind((conn.laddr[0], 0))
+                    except socket.error as err:
+                        if err.errno != errno.EADDRNOTAVAIL:
+                            raise
+            elif conn.family == AF_UNIX:
+                self.assertEqual(conn.status, psutil.CONN_NONE)
+
+        def check_type(conn):
+            # SOCK_SEQPACKET may happen in case of AF_UNIX socks
+            self.assertIn(conn.type, (SOCK_STREAM, SOCK_DGRAM, SOCK_SEQPACKET))
+            # if enum is not None:
+            #     assert isinstance(conn.type, enum.IntEnum), conn
+            # else:
+            #     assert isinstance(conn.type, int), conn
+            if conn.type == SOCK_DGRAM:
+                self.assertEqual(conn.status, psutil.CONN_NONE)
+
+        def check_addrs(conn):
+            # check IP address and port sanity
+            for addr in (conn.laddr, conn.raddr):
+                if conn.family in (AF_INET, AF_INET6):
+                    self.assertIsInstance(addr, tuple)
+                    if not addr:
+                        continue
+                    self.assertIsInstance(addr.port, int)
+                    assert 0 <= addr.port <= 65535, addr.port
+                    check_net_address(addr.ip, conn.family)
+                elif conn.family == AF_UNIX:
+                    self.assertIsInstance(addr, str)
+
+        def check_status(conn):
+            self.assertIsInstance(conn.status, str)
+            valids = [getattr(psutil, x) for x in dir(psutil)
+                      if x.startswith('CONN_')]
+            self.assertIn(conn.status, valids)
+
+        check_ntuple(conn)
+        check_fd(conn)
+        check_family(conn)
+        check_type(conn)
+        check_addrs(conn)
+        check_status(conn)
+
+
+# ---
+
 
 class TestFetchAll(Base, unittest.TestCase):
 
     def test_system(self):
         with create_sockets():
             for conn in psutil.net_connections('all'):
-                check_connection_ntuple(conn)
+                self.check_connection_ntuple(conn)
 
     def test_process(self):
         with create_sockets():
             for conn in psutil.Process().connections('all'):
-                check_connection_ntuple(conn)
+                self.check_connection_ntuple(conn)
 
 
 class TestUnconnectedSockets(Base, unittest.TestCase):
@@ -265,7 +358,7 @@ class TestConnectedSocket(Base, unittest.TestCase):
         def check_conn(proc, conn, family, type, laddr, raddr, status, kinds):
             all_kinds = ("all", "inet", "inet4", "inet6", "tcp", "tcp4",
                          "tcp6", "udp", "udp4", "udp6")
-            check_connection_ntuple(conn)
+            self.check_connection_ntuple(conn)
             self.assertEqual(conn.family, family)
             self.assertEqual(conn.type, type)
             self.assertEqual(conn.laddr, laddr)
@@ -429,7 +522,7 @@ class TestSystemWideConnections(Base, unittest.TestCase):
                 self.assertIn(conn.family, families, msg=conn)
                 if conn.family != AF_UNIX:
                     self.assertIn(conn.type, types_, msg=conn)
-                check_connection_ntuple(conn)
+                self.check_connection_ntuple(conn)
 
         with create_sockets():
             from psutil._common import conn_tmap
